requirements.txt
astroplan
astropy-base
botocore
boto3
click
healpy
matplotlib-base
numpy <2.4
pandas
psycopg2
rubin-nights >=0.10.0
rubin-scheduler =3.20.1
rubin-sim >=2.6.1

